Dedup NetworkSettingsBuilder
Move test to Core.
